feat: full patch lifecycle and graphql api

+++ bespoke-patches-server/src/patches/patches.resolver.ts
@@ -1,6 +1,8 @@
+import { ConfigService } from '@nestjs/config';
 import { Args, Mutation, Query, Resolver } from '@nestjs/graphql';
 import { ActionTokenService } from 'src/action-token/action-token.service';
-import { Patch } from './patch.model';
+import { MailService } from 'src/mail/mail.service';
+import { Patch, PatchOutput } from './patch.model';
 import { PatchesService } from './patches.service';
 import { UploadPatchInput } from './upload-patch.input';
 
@@ -9,14 +11,32 @@ export class PatchesResolver {
   constructor(
     private service: PatchesService,
     private actionTokenService: ActionTokenService,
+    private mailService: MailService,
+    private config: ConfigService,
   ) {}
 
-  @Query(() => Patch)
-  async patch(@Args('uuid') uuid: string) {
-    return null;
+  @Query(() => PatchOutput)
+  async patch(
+    @Args('uuid') uuid: string,
+    @Args('token', { nullable: true }) token: string | null,
+  ) {
+    return this.service.get(uuid, token);
   }
 
-  @Query(() => [Patch])
+  @Mutation(() => PatchOutput)
+  async moderatePatch(
+    @Args('uuid') uuid: string,
+    @Args('token', { nullable: true }) token: string | null,
+    @Args('approved') approved: boolean,
+  ) {
+    const patch = await this.service.moderate(uuid, token, approved);
+
+    this.mailService.sendModerationResult(patch, approved);
+
+    return patch;
+  }
+
+  @Query(() => [PatchOutput])
   async patches(
     @Args('tags', { defaultValue: [], type: () => [String] }) tags: string[],
     @Args('search', { defaultValue: null, type: () => String })
@@ -32,7 +52,10 @@ export class PatchesResolver {
     @Args('tokenUuid')
     tokenUuid: string,
   ) {
-    if (!(await this.actionTokenService.get(tokenUuid)).enabled) {
+    if (
+      !this.config.get<boolean>('DISABLE_ACTION_TOKEN_CHECK') &&
+      !(await this.actionTokenService.get(tokenUuid)).enabled
+    ) {
       throw new Error('Token not enabled');
     }
 
@@ -51,7 +74,11 @@ export class PatchesResolver {
 
     const patch = await this.service.save(patchToSave);
 
-    this.actionTokenService.deleteToken(tokenUuid);
+    if (!this.config.get<boolean>('DISABLE_ACTION_TOKEN_CHECK')) {
+      await this.actionTokenService.deleteToken(tokenUuid);
+    }
+    await this.mailService.sendSubmittedPatch(patch);
+
     return patch.uuid;
   }
 }
